newsletter-service/requirements.txt
requests==2.28.1
python-dotenv==1.0.0
mailjet-rest==1.3.0
feedparser==6.0.10
